feat & refactor: added test for tsp + renamed bitmasking.py #29
