+++ algorithms/dp/bitmasking_tsp.py
@@ -1,0 +1,15 @@
+"""
+Finds the length of the shortest Euler circuit using bitmasking and dynamic programming (DP).
+:param nodes: Contains integer tuples of the node coordinates.
+:returns: the length of the shortest Euler circuit.
+"""
+def tsp(nodes: list):
+    # TODO: implement tsp algorithm using bitmasking + dp
+    return -1
+
+# For debugging purposes
+def main():
+    print("")
+
+if __name__ == '__main__':
+    main()
